add data service & render to UI

+++ CryptoApp/Models/CoinModel.swift
@@ -2,7 +2,7 @@
 //  CoinModel.swift
 //  CryptoApp
 //
-//  Created by hanbiro on 10/17/23.
+//  Created by hanbiro - ANHDUC on 10/17/23.
 //
 
 import Foundation
@@ -59,7 +59,7 @@ struct CoinModel: Identifiable, Codable {
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
     let marketCapChangePercentage24H: Double?
-    let circulatingSupply, totalSupply, maxSupply, ath: Int?
+    let circulatingSupply, totalSupply, maxSupply, ath: Double?
     let athChangePercentage: Double?
     let athDate: String?
     let atl, atlChangePercentage: Double?
